requirements.txt
dash
dash-bootstrap-components
dash-core-components
futu-api
numpy
pandas
plotly
